# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ForumLite API Documentation",
        "description": "API documentation for the ForumLite application, providing endpoints for managing users, posts, and comments.",
        "contact": {
            "email": "eugenejrmartinez@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "ForumLite API Server"
        }
    ],
    "paths": {
...
        "/user/posts": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Get posts of the authenticated user",
                "description": "Returns a paginated list of posts created by the currently authenticated user.",
                "operationId": "getAuthenticatedUserPosts",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PostResource"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/comments": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Get comments of the authenticated user",
                "description": "Returns a paginated list of comments made by the currently authenticated user.",
                "operationId": "getAuthenticatedUserComments",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CommentResource"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationLinks": {
                "title": "Pagination Links",
                "description": "Links for paginated results",
                "properties": {
                    "first": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/api/posts?page=1",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/api/posts?page=10",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "format": "url",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/api/posts?page=2",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "Pagination Meta",
                "description": "Meta information for paginated results",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "path": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/api/posts"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "to": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 100
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "format": "url",
                                    "example": "http://localhost/api/posts?page=1",
                                    "nullable": true
                                },
                                "label": {
                                    "type": "string",
                                    "example": "&laquo; Previous"
                                },
                                "active": {
                                    "type": "boolean",
                                    "example": false
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "UnauthenticatedResponse": {
                "title": "Unauthenticated Response",
                "description": "Standard response for unauthenticated requests.",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "CommentResource": {
                "title": "Comment Resource",
                "description": "Comment resource representation",
                "properties": {
                    "id": {
                        "description": "Comment ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "content": {
                        "description": "Comment content",
                        "type": "string",
                        "example": "This is a great post!"
                    },
                    "created_at": {
                        "description": "Timestamp of comment creation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp of last comment update",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "post": {
                        "description": "Basic information about the post this comment belongs to",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "int64",
                                "example": 1
                            },
                            "title": {
                                "type": "string",
                                "example": "My First Post"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PostResource": {
                "title": "Post Resource",
                "description": "Post resource representation",
                "properties": {
                    "id": {
                        "description": "Post ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "description": "Post title",
                        "type": "string",
                        "example": "My First Post"
                    },
                    "content": {
                        "description": "Post content",
                        "type": "string",
                        "example": "This is the content of my first post."
                    },
                    "created_at": {
                        "description": "Timestamp of post creation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp of last post update",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "comments_count": {
                        "description": "Number of comments on the post",
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User Resource",
                "description": "User resource representation",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "User's name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "User's email address",
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "created_at": {
                        "description": "Timestamp of user creation",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Timestamp of last user update",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}